requirements.txt
langchain-pinecone
python-dotenv~=1.0.1
langchain
openai
tiktoken
PyFiglet